refactor shop and theme editors
