+++ apps/cms/src/app/cms/shop/[shop]/themes/ThemePreview.tsx
@@ -1,0 +1,19 @@
+"use client";
+
+import WizardPreview from "../../../wizard/WizardPreview";
+import type { CSSProperties } from "react";
+
+interface Props {
+  style: CSSProperties;
+  onTokenSelect: (token: string, coords?: { x: number; y: number }) => void;
+}
+
+export default function ThemePreview({ style, onTokenSelect }: Props) {
+  return (
+    <WizardPreview
+      style={style}
+      inspectMode
+      onTokenSelect={onTokenSelect}
+    />
+  );
+}
