+++ apps/cms/src/app/cms/shop/[shop]/themes/TypographySettings.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function TypographySettings() {
+  return null;
+}
